Noticed that this change made a significant improvement in serde::Serialize timings (on my machine at least). I'm surprised this worked so it might be good to see if this can be replicated elsewhere before it gets merged

After 431988f

```
    Finished bench [optimized] target(s) in 0.02s
     Running unittests src/lib.rs (target/release/deps/chrono-e1dcaaaa10e3f11a)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 177 filtered out; finished in 0.00s

     Running benches/chrono.rs (target/release/deps/chrono-ca1f633a26e467d4)
     Running benches/serde.rs (target/release/deps/serde-5c35e4d9face791b)
bench_ser_naivedatetime_writer
                        time:   [101.42 ns 101.76 ns 102.11 ns]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

bench_ser_naivedatetime_string
                        time:   [108.30 ns 108.59 ns 108.84 ns]

bench_ser_naivedate_writer
                        time:   [55.041 ns 55.269 ns 55.477 ns]

bench_ser_naivedate_string
                        time:   [60.872 ns 61.116 ns 61.352 ns]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```

After f26c2a9

```
   Compiling chrono v0.4.23 (/var/home/eric/Repos/chrono)
    Finished bench [optimized] target(s) in 3.53s
     Running unittests src/lib.rs (target/release/deps/chrono-e1dcaaaa10e3f11a)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 177 filtered out; finished in 0.00s

     Running benches/chrono.rs (target/release/deps/chrono-ca1f633a26e467d4)
     Running benches/serde.rs (target/release/deps/serde-5c35e4d9face791b)
bench_ser_naivedatetime_writer
                        time:   [76.851 ns 77.143 ns 77.406 ns]
                        change: [-25.180% -24.792% -24.424%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

bench_ser_naivedatetime_string
                        time:   [86.622 ns 86.961 ns 87.260 ns]
                        change: [-20.409% -19.997% -19.609%] (p = 0.00 < 0.05)
                        Performance has improved.

bench_ser_naivedate_writer
                        time:   [42.930 ns 43.114 ns 43.292 ns]
                        change: [-22.248% -21.797% -21.343%] (p = 0.00 < 0.05)
                        Performance has improved.

bench_ser_naivedate_string
                        time:   [51.183 ns 51.308 ns 51.424 ns]
                        change: [-16.972% -16.516% -16.079%] (p = 0.00 < 0.05)
                        Performance has improved.
```