@djc 2022-11-15
I agree that this is a very surprising result... Weird.