@pitdicker 2023-07-02
On my system the results are the opposite.

Before:
```
bench_ser_naivedatetime_writer
                        time:   [154.88 ns 155.64 ns 156.42 ns]
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

bench_ser_naivedatetime_string
                        time:   [188.50 ns 189.87 ns 191.22 ns]
Found 8 outliers among 100 measurements (8.00%)
  7 (7.00%) high mild
  1 (1.00%) high severe

bench_ser_naivedate_writer
                        time:   [89.863 ns 90.678 ns 91.418 ns]
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild

bench_ser_naivedate_string
                        time:   [106.47 ns 107.27 ns 108.08 ns]
Found 17 outliers among 100 measurements (17.00%)
  13 (13.00%) low mild
  2 (2.00%) high mild
  2 (2.00%) high severe
```

After:
```
bench_ser_naivedatetime_writer
                        time:   [163.83 ns 164.37 ns 164.96 ns]
                        change: [+5.2191% +5.8758% +6.5430%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe

bench_ser_naivedatetime_string
                        time:   [190.76 ns 192.26 ns 193.93 ns]
                        change: [+3.3698% +4.5953% +5.8575%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  7 (7.00%) high mild

bench_ser_naivedate_writer
                        time:   [94.864 ns 95.510 ns 96.200 ns]
                        change: [+2.1253% +2.9973% +3.9094%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

bench_ser_naivedate_string
                        time:   [112.50 ns 112.99 ns 113.50 ns]
                        change: [+3.8651% +4.6137% +5.3728%] (p = 0.00 < 0.05)
                        Performance has regressed.
```